Speed up login by trimming browser usage
Use Camoufox just for solving the Cloudflare challenge and then go back to plain requests to log in.
The login requests requires a Next.js 'next-action' header, which can be read from the loaded js files.
While loading the page and solving the Cloudflare challenge, we also get the js file and parse the correct 'next-action' hash to use in the subsequent requests.post login call.